fix: Cleanup the code
- Moved open to writeable so don't need to don't need to stick in Nav.
- Moved Modal to page instead of Nav
