+++ sk/src/lib/components/Nav.svelte
@@ -8,7 +8,7 @@
   import MagnifyingGlass from 'svelte-radix/MagnifyingGlass.svelte'
   import { toggleMode, ModeWatcher } from 'mode-watcher'
   import { Button } from '$lib/components/ui/button/index.js'
-  import AnilistModal from '$lib/components/AnilistModal.svelte'
+  import { metadata, openSearchModal } from '$lib/app/stores'
 
   import * as DropdownMenu from '$lib/components/ui/dropdown-menu'
   import * as Avatar from '$lib/components/ui/avatar'
@@ -34,35 +34,11 @@
     ['https://discord.com/invite/jPeeZewWRn', 'Discord'],
     ['https://sheet.releases.moe', 'Sheet']
   ]
+
 </script>
 
-<script lang='ts'>
-  import { onMount, onDestroy } from 'svelte';
-
-  export let search: boolean = false
-  export let ids: number[]
-  
-  let show: boolean
-  $: show = search
-
-  let open = false  
-
-  function handleKeyDown(event:KeyboardEvent) {
-    if (!show) return
-    if ((event.ctrlKey || event.metaKey) && event.key === 'f') {
-      event.preventDefault(); // Stop browser's default search
-      open = !open
-    }
-  }
-
-  onMount(() => {
-    window.addEventListener('keydown', handleKeyDown);
-  });
-
-  onDestroy(() => {
-    window.removeEventListener('keydown', handleKeyDown);
-  });
-
+<script lang='ts'>  
+  $: search = !!$metadata.title && $metadata.title !== 'Home'
 </script>
 
 <header class='sticky top-0 z-50 w-full border-b border-border/40 bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60'>
@@ -77,8 +53,8 @@
           {@const active = $page.url.pathname === path}
           <a href={`${base}${path}`} class='transition-colors hover:text-foreground/80 {active ? 'text-foreground' : 'text-foreground/60'}'>{label}</a>
         {/each}
-        {#if show}
-          <Button variant='ghost' size='icon' on:click={() => open=true}>
+        {#if search}
+          <Button variant='ghost' size='icon' on:click={() => $openSearchModal=true}>
             <MagnifyingGlass></MagnifyingGlass>
           </Button>  
         {/if}
@@ -128,5 +104,3 @@
     </div>
   </div>
 </header>
-
-<AnilistModal bind:open ids={ids} />
