+++ sk/src/routes/+layout.svelte
@@ -1,11 +1,12 @@
 <script context='module' lang='ts'>
-  import { metadata } from '$lib/app/stores'
+  import { metadata, openSearchModal } from '$lib/app/stores'
   import { ProgressBar } from '@prgm/sveltekit-progress-bar'
   import Nav from '$lib/components/Nav.svelte'
   import { site } from '$lib/config'
   import '../app.pcss'
   import { Toaster } from 'svelte-sonner'
   import { progress } from '$lib/components/entries/query'
+  import AnilistModal from '$lib/components/AnilistModal.svelte'
 </script>
 
 <script lang='ts'>
@@ -22,7 +23,8 @@
 </svelte:head>
 
 <ProgressBar zIndex={100} bind:this={$progress} />
-<Nav ids={$page.data.ids} search={!!$metadata.title && $metadata.title !== 'Home'} />
+<Nav/>
+<AnilistModal ids={$page.data.ids} bind:open={$openSearchModal} />
 <Toaster visibleToasts={3} position='top-right' theme='dark' richColors duration={10000} />
 
 <div class='container mx-auto py-5 flex-grow flex'>
